requirements.txt
fluvio==0.16.3
langchain_community==0.2.15
openai==1.43.0
pydantic==1.6.2
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.8
tavily_python==0.3.5
